Update Ex_1.3_fahr2celheading.c (#121)
Exercise 1.3 asks to modify the code given in the textbook to print a heading above the table. The version I am proposing contains a few changes that can make the table look better. The table header would include the temperature types to distinguish between the two columns. "printf" within the while loop allows the table to be organized better aligning it on the right. 

Thank you.